requirements.txt
python-binance==1.0.19
pandas==2.1.0
python-dateutil==2.8.2
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
plotly==5.18.0
dash==2.14.2
fastapi==0.104.1
uvicorn==0.24.0
flask==3.0.0
flask-cors==4.0.0 